FireWall
In the Windows XP operating system, the firewall program was developed
on VC6.0 platform, with the help of Windows DDK 7600.16385.1 and MSDN.
The firewall provided users with a friendly user interface. After the
actual test, it ran stably, blocked the data of a specific IP address in
order to filter the malicious web site, but also to a complete work on
all TCP packets, UCP packets, ICMP packets interception and release.